refactor(scope): split ad_completeness_review below the 800-line cap
The pre-commit ``enforce 800-line limit`` hook blocks CI on
``fix/task-outcome-invariant`` because two files crept past 800 lines
during the D1–D6 design fix:

  * ``app/ai/stop_gates/ad_completeness_review.py`` — 960 lines
  * ``app/skills_v2/amazon-ads/scripts/ads_bulk.py`` — 803 lines

This is a behaviour-preserving refactor purely to clear the hook.

The new module ``app/ai/stop_gates/ad_completeness_campaign_blocks.py``
owns the per-campaign reconciliation contract:

  * the campaign-heading regex (``_CAMPAIGN_HEAD_RE``, re-exported for
    the duplicate-id scan in ``check``);
  * the per-campaign search-term reconciliation regexes
    (``_RECONCILE_RE``, ``_NO_SEARCHTERM_RE``, ``_PENDING_WORK_RE``);
  * the spend-tolerance helpers (``_has_valid_escape``, ``_num``,
    ``_within``);
  * ``_check_campaign_blocks`` itself (146 lines of D2/D3 logic).

``ad_completeness_review.py`` keeps ``GATE_NAME``, the stall-state
dicts, ``is_stalled``, ``drill_incomplete_reason``, the combo-header /
progress / drill-table / defer / garbled regexes, and ``check``. Public
API (``drill_incomplete_reason`` as called from
``app/ai/bash_safety.py``) is unchanged.

``ads_bulk.py``: ``cmd_scope`` docstring consolidated from 11 → 8 lines
to land the file at exactly 800.

Verification:

  * ``pre-commit run --files …`` — all hooks pass (including the
    800-line hook)
  * ``pytest tests/unit/test_stop_gates.py`` — 177 / 177 pass (same as
    pre-split)
  * ``pytest tests/workflow …`` (skipping the two
    ``openpyxl``/py3.13-deprecation-affected suites) — 589 / 589 pass

Diff data-safety: no real brand / store / SKU / ASIN / campaign-id
tokens; only generic shapes (``A0EXAMPLE…``, ``C_DEMO0001``, regex
patterns).

via [Happy](https://happy.engineering)
